03_backend_developement/06_mongo_basics/d07c1d6a6741450f3369f73957d8b6a0f6764bca/03_data_validation/src/createBooksCollection.ts
import * as mongo from "mongodb";

export const booksValidator = {
  validator: {
    $jsonSchema: {
      bsonType: "object",
      additionalProperties: false,
      required: ["_id", "title", "author", "description"],
      properties: {
        _id: {
          bsonType: "objectId",
          description: "must be an objectId and is required",
        },
        title: {
          bsonType: "string",
          description: "must be a string and is required",
        },
        author: {
          bsonType: "string",
          description: "must be a string and is required",
        },
        description: {
          bsonType: "string",
          maxLength: 100,
          description: "must be an string < 100 characters and is required",
        },
      },
    },
  },
};

export function createBooksCollection(db: mongo.Db): Promise<mongo.Collection> {
  return db.createCollection("books", booksValidator);
}
